[Refactor] Change some type from DateTimeValue to VecDateTimeValue (#266)

+++ be/src/vec/core/types.h
@@ -220,8 +220,8 @@ struct TypeId<Int128> {
     static constexpr const TypeIndex value = TypeIndex::Int128;
 };
 
-using Date = Int128;
-using DateTime = Int128;
+using Date = Int64;
+using DateTime = Int64;
 
 /// Own FieldType for Decimal.
 /// It is only a "storage" for decimal. To perform operations, you also have to provide a scale (number of digits after point).
